backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.3.0

# Gemini LLM
google-generativeai==0.7.2

# Vector DB
chromadb==0.5.0

# Telegram Bot
python-telegram-bot==21.3

# Scheduling
apscheduler==3.10.4

# Trend Analysis
pytrends==4.9.2
requests==2.32.3

# PDF Export
reportlab==4.2.2

# Async HTTP
httpx==0.27.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
